Implement finite damage zone

+++ Assets/Scripts/PlayerDamageZone.cs
@@ -3,13 +3,18 @@ using System.Collections.Generic;
 using UnityEngine;
 
 [RequireComponent(typeof(Collider2D))]
-public class PlayerDamageZone : MonoBehaviour {	
+public class PlayerDamageZone : MonoBehaviour
+{
+	public bool Kill;
+	public float Damage;
+	
 	private void OnTriggerEnter2D(Collider2D other)
 	{
 		var player = other.GetComponent<Player>();
 		if (player != null)
 		{
-			Player.Kill(null, player);
+			if (Kill) Player.Kill(null, player);
+			else      Player.DealDamage(null, player, Damage);
 		}
 	}
 }
